Split image into bootstrap and kernel images.
Due to issues with the linker, namely ld not being able to
perform the relocation for "jmp kmain" if kmain's address
is above 32bit, the bootstrap and the kernel are now a 32bit elf
and a flat 64bit binary respectively.

boot.bin clears the bss for kernel.bin, and jumps to kmain.
kernel.bin now contains a small header in the first 24 bytes
which contains the bss offset, length and address of kmain.

The long mode page tables are contained in the first 20k of the BSS,
accounted for by kernel.bin's linker script.